requirements.txt
arabic-reshaper==2.1.1
asgiref==3.2.10
certifi==2020.11.8
chardet==3.0.4
cloudinary==1.22.0
colorama==0.4.4
decorate-url==1.1
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.2
django-admin-honeypot==1.1.0
django-ckeditor==6.0.0
django-cloudinary-storage==0.3.0
django-countries==7.0
django-crispy-forms==1.9.2
django-decorator-include==3.0
django-google-translate==1.1
django-heroku==0.3.1
django-honeypot-signals==0.1.1
django-import-export==2.4.0
django-js-asset==1.2.2
django-modeladmin-reorder==0.3.1
django-recaptcha==2.0.6
django-registration-redux==2.8
django-social-share==2.1.0
et-xmlfile==1.0.1
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
idna==2.10
ipcalc==1.99.0
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.5
Pillow==8.0.0
psycopg2==2.8.6
PyPDF2==1.26.0
python-bidi==0.4.2
python-dotenv==0.15.0
python-magic==0.4.18
pytz==2020.1
PyYAML==5.3.1
reportlab==3.5.65
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
tablib==2.0.0
urllib3==1.26.2
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.5
xlrd==1.2.0
xlwt==1.3.0
